callAI/requirements.txt
fastapi 
uvicorn
twilio
python-multipart
websockets
openai-whisper
numpy
soundfile
pywav
scipy
pymongo[srv]
mongoengine
gtts
google.generativeai
pymupdf
python-dotenv